minor: move sequence_iterator.py to utilities
