+++ fedot/core/utilities/sequence_iterator.py
