+++ test/unit/optimizer/test_iterator.py
@@ -1,4 +1,4 @@
-from fedot.core.optimisers.gp_comp.iterator import SequenceIterator, fibonacci_sequence
+from fedot.core.utilities.sequence_iterator import SequenceIterator, fibonacci_sequence
 
 
 def test_iterator_without_constraints():
